17/admin crud entry (#26)
* 17: Добавить роуты для crud операций

* 17: Дописать контроллеры для crud операций

* 17: Дописать поля для crud операций

* 17: Дописать базовые шаблоны админки

* 17: Добавить валидацию полей.

* 17: Сгруппировать роуты и подправить шаблон меню

* 17: Исправить crud контроллеры.

* 17: Фиксы.

- Расширить сетку до 1440+
- фикс меню
- удалить неиспользуемые иконки
- сборка нового фронта

* 17: Добавить автоматическое создание slug, если поле пустое.

* 17: Сборка фронта.

* 17: Добавить логи

* 17: Подправить шаблоны админки с учетом ширины вывода логов

* 17: Сборка фронта

* 17: fix. Исправить шаблоны

* 17: fix. Исправить контроллеры.

+++ app/Http/Controllers/Admin/CrudController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\RedirectResponse;
 
 class CrudController extends Controller
 {
@@ -52,12 +53,20 @@ class CrudController extends Controller
      */
     public function newEntry(): View
     {
-        $fields = $this->all_fields->merge($this->update_fields);
+        $fields = $this->all_fields->merge($this->create_fields);
         return view('admin.base.new_item', ['fields' => $fields, 'routes' => $this->routes]);
     }
 
+    /**
+     * @param Request $request
+     * @return RedirectResponse
+     * create new entry
+     */
     public function create(Request $request): RedirectResponse
     {
+        if (!$this->buttons['add']) {
+            return redirect()->route('home');
+        }
         $this->model::create($request->all());
         return redirect()->route($this->routes['all']);
     }
@@ -80,8 +89,11 @@ class CrudController extends Controller
      * @return RedirectResponse
      * update entry properties
      */
-    public function save(Request $request, $id): RedirectResponse
+    public function update(Request $request, $id): RedirectResponse
     {
+        if (!$this->buttons['edit']) {
+            return redirect()->route('home');
+        }
         $this->model::find($id)->update($request->all());
         return redirect()->route($this->routes['all']);
     }
@@ -93,6 +105,9 @@ class CrudController extends Controller
      */
     public function delete($id): RedirectResponse
     {
+        if (!$this->buttons['delete']) {
+            return redirect()->route('home');
+        }
         $this->model::find($id)->delete();
         return redirect()->route($this->routes['all']);
     }
@@ -197,33 +212,17 @@ class CrudController extends Controller
     }
 
     /**
-     * @return JSON string data
-     * @param string $mark
+     * @param string|null $mark
+     * @return string
      */
-    public function getJson($mark = null) 
+    public function getJson(string $mark = null): string
     {
-        if ($mark === 'home') {
-            return json_encode(
-                [
-                    'entries' => collect([]),
-                    'columns' => $this->columns,
-                    'title' => $this->title,
-                    'buttons' => $this->buttons,
-                    'routes' => $this->routes,
-                ]
-            );
-        } else {
-            return json_encode(
-                [
-                    'entries' => $this->model::all(),
-                    'columns' => $this->columns,
-                    'title' => $this->title,
-                    'buttons' => $this->buttons,
-                    'routes' => $this->routes,
-                ]
-            );
-        }
-        
-        
+        return collect([
+            'entries' => $mark === 'home' ? collect([]) : $this->model::all(),
+            'columns' => $this->columns,
+            'title' => $this->title,
+            'buttons' => $this->buttons,
+            'routes' => $this->routes,
+        ])->toJson();
     }
 }
